requirements.txt
streamlit>=1.32.0
openai>=1.12.0
pydantic>=2.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
